+++ data/scripts/TEST_update.py
@@ -1,0 +1,17 @@
+# Script for test
+
+def update(plc):
+    plc.d[0] = -1024
+    plc.d[1] = 200
+    plc.d[3] = 17217
+    plc.d[4] = 16964
+    plc.d[5] = 17217
+    plc.d[6] = 16964
+    plc.d[7] = 17217
+    plc.d[8] = 16964
+    plc.d[9] = 17217
+    plc.d[10] = 16964
+    plc.d[11] = 17217
+    plc.d[12] = 16964
+    plc.d[13] = 17217
+    pass