+++ data/scripts/ww_update.py
@@ -1,0 +1,4 @@
+# Script for ww
+
+def update(plc):
+    pass